requirements.txt
pythonnet==3.0.0a1
pysqlite3>=0.4.7
pywebview>=3.6.2
requests>=2.21.0
pygame>=2.1.0
